+++ R/plotly_data.R
@@ -88,16 +88,19 @@ print.plotly_data <- function(x, ...) {
   x
 }
 
-#' Shared data
+#' Highlight/query data based on primary key
 #' 
-#' This is simply a wrapper around `crosstalk::SharedData$new()` to make it easier
-#' to use and explain conceptually. It also makes it more discoverable if one
+#' This function simply creates an object of class [crosstalk::SharedData].
+#' The reason it exists is to make it easier to teach others how to leverage
+#' it's functionality in plotly. It also makes it more discoverable if one
 #' is already aware of [highlight].
 #' 
-#' @param ... arguments passed to crosstalk::SharedData$new()
+#' @param ... arguments passed to `crosstalk::SharedData$new()`
 #' @export
+#' @author Carson Sievert
 #' @return An object of class [crosstalk::SharedData]
-highlight_unit <- function(...) {
+#' @seealso [highlight]
+highlight_key <- function(...) {
   crosstalk::SharedData$new(...)
 }