+++ demo/crosstalk-filter-lines.R
@@ -3,7 +3,7 @@ library(ggplot2)
 library(gapminder)
 library(plotly)
 
-sd <- highlight_unit(gapminder)
+sd <- highlight_key(gapminder)
 
 g <- ggplot(sd, aes(year, lifeExp, color = country, group = country)) +
   geom_line()