+++ demo/crosstalk-highlight-ggpairs.R
@@ -1,5 +1,5 @@
 library(GGally)
-d <- highlight_unit(iris)
+d <- highlight_key(iris)
 p <- ggpairs(d, aes(colour = Species), columns = 1:5)
 ggplotly(p) %>% 
   highlight("plotly_selected")