+++ demo/crosstalk-highlight-subplot.R
@@ -1,6 +1,6 @@
 library(plotly)
 
-d <- highlight_unit(mtcars)
+d <- highlight_key(mtcars)
 s <- subplot(
   qplot(data = d, x = mpg, y = wt),
   qplot(data = d, x = mpg, y = vs)